% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_zero.R
\name{drop_zero}
\alias{drop_zero}
\title{Drop Zero Count Elements}
\usage{
drop_zero(x, ...)
}
\arguments{
\item{x}{A \code{\link[synonym]{get_synonym}} object.}

\item{\ldots}{ignored.}
}
\value{
Returns a list with \code{NA} elements removed.
}
\description{
The \code{\link[synonym]{get_synonym}} terms that are found in the key but
that do not match the relevant distance return an \code{NA}.  This function
conveniently drops these elements.
}
\examples{
get_synonym(c('cat', 'dog', 'chicken', 'dfsf'))

drop_zero(
    get_synonym(c('cat', 'dog', 'chicken', 'dfsf'))
)
}

