% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_synonym.R
\name{get_synonym}
\alias{get_synonym}
\title{Get Synonyms}
\usage{
get_synonym(x, distance = 1, key = synonym::synonym_key, ...)
}
\arguments{
\item{x}{A character vector of terms to find synonyms for.}

\item{distance}{A relevance distance 1-most relevant, 2-mid relevancy, &
3-furthest relevance.  This value can be given as a single integer element
or as a vector of integers.}

\item{key}{A synonym key consisting of a \code{term} column and a \code{json}
column.  The \code{\link[synonym]{synonym_key}} is used by default.}

\item{\ldots}{Currently ignored.}
}
\value{
Returns a \code{\link[base]{list}} of \code{\link[base]{vector}}s of
relevant synonyms.  The \code{\link[base]{list}} is equal or less than the
length of \code{x}.
}
\description{
Look up synonyms for term(s) with relevant distance filtering.
}
\note{
\code{x} terms not found in \code{key} are not included in the output.
\code{x} terms found in \code{key} but not meeting the distance criteria
return as \code{NA}.
}
\examples{
get_synonym(c('cat', 'dog', 'chicken', 'dfsf'))

get_synonym('cat', 1)
get_synonym('cat', 2)
get_synonym('cat', 3)
get_synonym('cat', 2:3)

drop_zero(
    get_synonym(c('cat', 'dog', 'chicken', 'dfsf'))
)
}
\keyword{synonym}

