% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_synonym.R
\name{print.get_synonym}
\alias{print.get_synonym}
\title{Prints a get_synonym Object}
\usage{
\method{print}{get_synonym}(x, ...)
}
\arguments{
\item{x}{A get_synonym object.}

\item{\ldots}{ignored.}
}
\description{
Prints a get_synonym object
}
